requirements.txt
openai==0.27.6
pytest==7.3.1
python-dotenv==1.0.0
uvicorn==0.22.0
tiktoken==0.3.3
fastapi==0.95.2
pydantic==1.10.7
pinecone-io==0.0.1
